+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-# dev-master (2019-03-14)
+# dev-master (2019-03-16)
+
+# 0.0.3 (2019-03-16)
+
+- Added group filters.
+- Added expressions for filters.
+- Added batch query.
+- Updated unitTest.
 
 # 0.0.2 (2019-03-14)